requirements.txt
click>=8.1.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
pinecone>=7.0.0
tiktoken>=0.5.1
rich>=13.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.3 